[nfc] Break Pipe code out of Valid.scala
Move code related to the `Pipe` utility into its own file.  Previously,
this code lived inside Valid.scala which would make it harder to locate.
